feat: add flag to disable noVNC

+++ scripts/init.sh
@@ -11,14 +11,16 @@ pacman -Syu --noconfirm xorg-server \
 	i3-wm python-setuptools ttf-dejavu
 
 # Install noVNC
-wget https://github.com/novnc/websockify/archive/v${websockify_version}.tar.gz -O /websockify.tar.gz \
-	&& tar -xvf /websockify.tar.gz -C / \
-	&& cd /websockify-${websockify_version} \
-	&& python setup.py install \
-	&& cd / && rm -r /websockify.tar.gz /websockify-${websockify_version} \
-	&& wget https://github.com/novnc/noVNC/archive/v${noVNC_version}.tar.gz -O /noVNC.tar.gz \
-	&& tar -xvf /noVNC.tar.gz -C / \
-	&& mv /noVNC-${noVNC_version} /noVNC \
-	&& cd /noVNC \
-	&& ln -s vnc.html index.html \
-	&& rm /noVNC.tar.gz
+if [ "$DISABLE_NOVNC" != "true" ]; then
+	wget https://github.com/novnc/websockify/archive/v${websockify_version}.tar.gz -O /websockify.tar.gz \
+		&& tar -xvf /websockify.tar.gz -C / \
+		&& cd /websockify-${websockify_version} \
+		&& python setup.py install \
+		&& cd / && rm -r /websockify.tar.gz /websockify-${websockify_version} \
+		&& wget https://github.com/novnc/noVNC/archive/v${noVNC_version}.tar.gz -O /noVNC.tar.gz \
+		&& tar -xvf /noVNC.tar.gz -C / \
+		&& mv /noVNC-${noVNC_version} /noVNC \
+		&& cd /noVNC \
+		&& ln -s vnc.html index.html \
+		&& rm /noVNC.tar.gz
+fi
